+++ notifications/subscriptions/raw_client.go
@@ -6,11 +6,11 @@ import (
 	context "context"
 	http "net/http"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v24"
-	core "github.com/KardFinancial/kard-go-sdk/v24/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v24/internal"
-	notifications "github.com/KardFinancial/kard-go-sdk/v24/notifications"
-	option "github.com/KardFinancial/kard-go-sdk/v24/option"
+	kard "github.com/KardFinancial/kard-go-sdk/v25"
+	core "github.com/KardFinancial/kard-go-sdk/v25/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v25/internal"
+	notifications "github.com/KardFinancial/kard-go-sdk/v25/notifications"
+	option "github.com/KardFinancial/kard-go-sdk/v25/option"
 )
 
 type RawClient struct {