+++ organizations/placements/organizations_placements_test/organizations_placements_test.go
@@ -10,9 +10,9 @@ import (
 	os "os"
 	testing "testing"
 
-	client "github.com/KardFinancial/kard-go-sdk/v24/client"
-	option "github.com/KardFinancial/kard-go-sdk/v24/option"
-	organizations "github.com/KardFinancial/kard-go-sdk/v24/organizations"
+	client "github.com/KardFinancial/kard-go-sdk/v25/client"
+	option "github.com/KardFinancial/kard-go-sdk/v25/option"
+	organizations "github.com/KardFinancial/kard-go-sdk/v25/organizations"
 	require "github.com/stretchr/testify/require"
 )