+++ users/rewards/users_rewards_test/users_rewards_test.go
@@ -10,10 +10,10 @@ import (
 	os "os"
 	testing "testing"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v24"
-	client "github.com/KardFinancial/kard-go-sdk/v24/client"
-	option "github.com/KardFinancial/kard-go-sdk/v24/option"
-	users "github.com/KardFinancial/kard-go-sdk/v24/users"
+	kard "github.com/KardFinancial/kard-go-sdk/v25"
+	client "github.com/KardFinancial/kard-go-sdk/v25/client"
+	option "github.com/KardFinancial/kard-go-sdk/v25/option"
+	users "github.com/KardFinancial/kard-go-sdk/v25/users"
 	require "github.com/stretchr/testify/require"
 )